refactor(analysis): rename parserdialog for all tools (#4008)
