+++ analysis/analysers/tools/InspectionTool/src/main/kotlin/de/maibornwolff/codecharta/analysers/tools/inspection/Dialog.kt
@@ -8,7 +8,7 @@ import de.maibornwolff.codecharta.dialogProvider.myPromptDefaultFileFolderInput
 import de.maibornwolff.codecharta.dialogProvider.myPromptInputNumber
 import de.maibornwolff.codecharta.serialization.FileExtension
 
-class ParserDialog {
+class Dialog {
     companion object : ParserDialogInterface {
         override fun collectParserArgs(session: Session): List<String> {
             val inputFileName: String = session.myPromptDefaultFileFolderInput(
