+++ analysis/analysers/tools/InspectionTool/src/main/kotlin/de/maibornwolff/codecharta/analysers/tools/inspection/InspectionTool.kt
@@ -78,7 +78,7 @@ class InspectionTool(
         }
     }
 
-    override fun getDialog(): ParserDialogInterface = ParserDialog
+    override fun getDialog(): ParserDialogInterface = Dialog
 
     override fun isApplicable(resourceToBeParsed: String): Boolean {
         return false
