+++ analysis/analysers/tools/ValidationTool/src/main/kotlin/de/maibornwolff/codecharta/analysers/tools/validation/ValidationTool.kt
@@ -41,7 +41,7 @@ class ValidationTool : Callable<Unit?>, AnalyserInterface {
         return null
     }
 
-    override fun getDialog(): ParserDialogInterface = ParserDialog
+    override fun getDialog(): ParserDialogInterface = Dialog
 
     override fun isApplicable(resourceToBeParsed: String): Boolean {
         return false
